Fix a regression with accessing the Flutter store file.
