+++ packages/devtools_shared/CHANGELOG.md
@@ -2,6 +2,7 @@
 * Added helper `deserialize` and `deserializeNullable`
 * Extended serialization for `HeapSample` and `ExtensionEvents`
 * Added mixin `Serializable`
+* Fix a regression with accessing the Flutter store file.
 
 # 10.0.0-dev.2
 * Support detecting package roots for nested Dart projects in the
