+++ packages/devtools_shared/lib/src/server/file_system.dart
@@ -7,7 +7,7 @@ import 'dart:io';
 
 import 'package:path/path.dart' as path;
 
-import 'usage.dart';
+import 'devtools_store.dart';
 
 // ignore: avoid_classes_with_only_static_members, requires refactor.
 class LocalFileSystem {
@@ -76,9 +76,95 @@ class LocalFileSystem {
     return jsonEncode(json);
   }
 
+  /// The location of the Flutter store file, ~/.flutter.
+  static String flutterStoreLocation() {
+    return path.join(_userHomeDir(), '.flutter');
+  }
+
   /// Whether the flutter store file exists.
   static bool flutterStoreExists() {
-    final flutterStore = File(path.join(_userHomeDir(), '.flutter'));
+    final flutterStore = File(flutterStoreLocation());
     return flutterStore.existsSync();
   }
 }
+
+class IOPersistentProperties extends PersistentProperties {
+  IOPersistentProperties(
+    String name, {
+    String? documentDirPath,
+  }) : super(name) {
+    final String fileName = name.replaceAll(' ', '_');
+    documentDirPath ??= LocalFileSystem._userHomeDir();
+    _file = File(path.join(documentDirPath, fileName));
+    if (!_file.existsSync()) {
+      _file.createSync(recursive: true);
+    }
+    syncSettings();
+  }
+
+  IOPersistentProperties.fromFile(File file) : super(path.basename(file.path)) {
+    _file = file;
+    if (!_file.existsSync()) {
+      _file.createSync(recursive: true);
+    }
+    syncSettings();
+  }
+
+  late File _file;
+
+  late Map<String, Object?> _map;
+
+  @override
+  // ignore: avoid-dynamic, necessary here.
+  dynamic operator [](String key) => _map[key];
+
+  @override
+  void operator []=(String key, Object? value) {
+    if (value == null && !_map.containsKey(key)) return;
+    if (_map[key] == value) return;
+
+    if (value == null) {
+      _map.remove(key);
+    } else {
+      _map[key] = value;
+    }
+
+    try {
+      _file.writeAsStringSync('${_jsonEncoder.convert(_map)}\n');
+    } catch (_) {}
+  }
+
+  @override
+  void syncSettings() {
+    try {
+      String contents = _file.readAsStringSync();
+      if (contents.isEmpty) contents = '{}';
+      _map = (jsonDecode(contents) as Map).cast<String, Object>();
+    } catch (_) {
+      _map = {};
+    }
+  }
+
+  void remove(String propertyName) {
+    _map.remove(propertyName);
+  }
+}
+
+abstract class PersistentProperties {
+  PersistentProperties(this.name);
+
+  final String name;
+
+  // ignore: avoid-dynamic, dynamic by design.
+  dynamic operator [](String key);
+
+  // ignore: avoid-dynamic, dynamic by design.
+  void operator []=(String key, dynamic value);
+
+  /// Re-read settings from the backing store.
+  ///
+  /// May be a no-op on some platforms.
+  void syncSettings();
+}
+
+const JsonEncoder _jsonEncoder = JsonEncoder.withIndent('  ');
